reworking directory structure
