+++ quadratic-core/src/controller/execution/mod.rs
@@ -1,0 +1,3 @@
+/// This module handles the application of operations to the Grid.
+///
+pub mod execute_operation;
