+++ quadratic-core/src/controller/export.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use csv::Writer;
 
-use super::{auto_complete::cell_values_in_rect, GridController};
+use super::GridController;
 use crate::{grid::SheetId, Rect};
 
 impl GridController {
@@ -10,7 +10,8 @@ impl GridController {
     /// Returns a [`String`].
     pub fn export_csv_selection(&self, sheet_id: SheetId, selection: &Rect) -> Result<String> {
         let sheet = self.sheet(sheet_id);
-        let values = cell_values_in_rect(selection, sheet)?
+        let values = sheet
+            .cell_values_in_rect(selection)?
             .into_cell_values_vec()
             .iter()
             .map(|record| record.to_string())
