+++ quadratic-core/src/controller/user_actions/mod.rs
@@ -1,0 +1,8 @@
+/// These are all user-initiated actions on the grid.
+///
+pub mod auto_complete;
+pub mod borders;
+pub mod cells;
+pub mod formatting;
+pub mod import;
+pub mod undo;
