+++ quadratic-core/src/grid/mod.rs
@@ -13,7 +13,7 @@ mod bounds;
 mod code;
 mod column;
 pub mod file;
-mod formatting;
+pub mod formatting;
 mod ids;
 pub mod js_types;
 mod offsets;
