+++ quadratic-multiplayer/src/file.rs
@@ -1,6 +1,8 @@
 use anyhow::Result;
 use quadratic_core::{
-    controller::{operation::Operation, transaction_summary::TransactionSummary, GridController},
+    controller::{
+        operations::operation::Operation, transaction_summary::TransactionSummary, GridController,
+    },
     grid::{file::import, Grid},
 };
